+++ common/src/main/java/net/modgarden/barricade/data/BlockedDirections.java
@@ -17,17 +17,17 @@ import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.List;
 
-public record BlockedDirectionsComponent(EnumSet<Direction> directions) {
-    public static final Codec<BlockedDirectionsComponent> CODEC = StringRepresentable.fromEnum(Direction::values).listOf().flatXmap(directions -> {
+public record BlockedDirections(EnumSet<Direction> directions) {
+    public static final Codec<BlockedDirections> CODEC = StringRepresentable.fromEnum(Direction::values).listOf().flatXmap(directions -> {
         if (directions.isEmpty()) {
             return DataResult.error(() -> "Must specify at least one direction.");
         }
-        return DataResult.success(BlockedDirectionsComponent.of(directions.toArray(Direction[]::new)));
+        return DataResult.success(BlockedDirections.of(directions.toArray(Direction[]::new)));
     }, component -> DataResult.success(List.copyOf(component.directions)));
-    public static final StreamCodec<ByteBuf, BlockedDirectionsComponent> STREAM_CODEC = ByteBufCodecs.fromCodec(CODEC);
+    public static final StreamCodec<ByteBuf, BlockedDirections> STREAM_CODEC = ByteBufCodecs.fromCodec(CODEC);
 
-    public static BlockedDirectionsComponent of(Direction... directions) {
-        return new BlockedDirectionsComponent(EnumSet.copyOf(Arrays.stream(directions).toList()));
+    public static BlockedDirections of(Direction... directions) {
+        return new BlockedDirections(EnumSet.copyOf(Arrays.stream(directions).toList()));
     }
 
     public boolean shouldBlock(BlockPos pos, CollisionContext context) {
@@ -71,7 +71,7 @@ public record BlockedDirectionsComponent(EnumSet<Direction> directions) {
     public boolean equals(Object obj) {
         if (obj == this)
             return true;
-        if (!(obj instanceof BlockedDirectionsComponent other))
+        if (!(obj instanceof BlockedDirections other))
             return false;
         return other.directions.equals(directions);
     }